moved some rules around and used require
