+++ recommended.js
@@ -1,0 +1,158 @@
+const baseRules = {
+  'array-callback-return': 1,
+  'arrow-body-style': [1, 'as-needed'],
+  camelcase: [1, { properties: 'never' }],
+  'comma-dangle': [
+    2,
+    {
+      arrays: 'ignore',
+      objects: 'ignore',
+      imports: 'ignore',
+      exports: 'ignore',
+      functions: 'never',
+    },
+  ],
+  complexity: [1, 20],
+  'constructor-super': 1,
+  curly: [1, 'multi-line'],
+  'default-case': [1, { commentPattern: '^no default$' }],
+  'eol-last': 2,
+  eqeqeq: [1, 'allow-null'],
+  'guard-for-in': 1,
+  'linebreak-style': [2, 'unix'],
+  'max-classes-per-file': [1, 4],
+  'max-depth': [1, 4],
+  'max-lines': [1, 400],
+  'max-lines-per-function': [
+    1,
+    { max: 150, skipBlankLines: true, skipComments: true, IIFEs: true },
+  ],
+  'max-nested-callbacks': [1, 4],
+  'max-params': [1, 4],
+  'max-statements': [1, 40],
+  'new-cap': [2, { newIsCap: true, capIsNew: false }],
+  'new-parens': 1,
+  'no-array-constructor': 1,
+  'no-caller': 1,
+  'no-class-assign': 1,
+  'no-cond-assign': [1, 'always'],
+  'no-console': [1],
+  'no-const-assign': 1,
+  'no-constant-condition': [1, { checkLoops: false }],
+  'no-control-regex': 1,
+  'no-debugger': 1,
+  'no-delete-var': 1,
+  'no-dupe-args': 1,
+  'no-dupe-class-members': 1,
+  'no-dupe-keys': 1,
+  'no-duplicate-case': 1,
+  'no-duplicate-imports': 2,
+  'no-empty-character-class': 1,
+  'no-empty-pattern': 1,
+  'no-eval': 2,
+  'no-ex-assign': 1,
+  'no-extend-native': 2,
+  'no-extra-bind': 1,
+  'no-extra-label': 1,
+  'no-extra-parens': [1, 'functions'],
+  'no-fallthrough': 1,
+  'no-func-assign': 1,
+  'no-global-assign': 2,
+  'no-implied-eval': 1,
+  'no-invalid-regexp': 1,
+  'no-irregular-whitespace': 1,
+  'no-iterator': 1,
+  'no-label-var': 1,
+  'no-labels': [1, { allowLoop: false, allowSwitch: false }],
+  'no-lone-blocks': 1,
+  'no-loop-func': 1,
+  'no-mixed-operators': [
+    1,
+    {
+      groups: [
+        ['&', '|', '^', '~', '<<', '>>', '>>>'],
+        ['==', '!=', '===', '!==', '>', '>=', '<', '<='],
+        ['&&', '||'],
+        ['in', 'instanceof'],
+      ],
+      allowSamePrecedence: false,
+    },
+  ],
+  'no-multi-str': 1,
+  'no-native-reassign': 1,
+  'no-negated-in-lhs': 1,
+  'no-new-func': 1,
+  'no-new-object': 1,
+  'no-new-require': 1,
+  'no-new-symbol': 1,
+  'no-new-wrappers': 1,
+  'no-obj-calls': 1,
+  'no-octal': 1,
+  'no-octal-escape': 1,
+  'no-path-concat': 1,
+  'no-proto': 1,
+  'no-redeclare': 1,
+  'no-regex-spaces': 1,
+  'no-restricted-syntax': [1, 'LabeledStatement', 'WithStatement'],
+  'no-return-assign': [1, 'except-parens'],
+  'no-script-url': 1,
+  'no-self-assign': 1,
+  'no-self-compare': 1,
+  'no-sequences': 1,
+  'no-shadow-restricted-names': 1,
+  'no-sparse-arrays': 1,
+  'no-template-curly-in-string': 1,
+  'no-this-before-super': 1,
+  'no-throw-literal': 1,
+  'no-trailing-spaces': 1,
+  'no-undef': 2,
+  'no-undef-init': 2,
+  'no-unexpected-multiline': 1,
+  'no-unneeded-ternary': [1, { defaultAssignment: false }],
+  'no-unreachable': 1,
+  'no-unsafe-finally': 1,
+  'no-unsafe-negation': 1,
+  'no-unused-expressions': [1, { allowTernary: true, allowShortCircuit: true }],
+  'no-unused-labels': 1,
+  'no-unused-vars': [
+    1,
+    {
+      vars: 'local',
+      varsIgnorePattern: '^_',
+      argsIgnorePattern: '^_|...props|props',
+      args: 'after-used',
+      ignoreRestSiblings: true,
+    },
+  ],
+  'no-use-before-define': [1, 'nofunc'],
+  'no-useless-call': 1,
+  'no-useless-computed-key': 1,
+  'no-useless-concat': 1,
+  'no-useless-constructor': 1,
+  'no-useless-escape': 1,
+  'no-useless-rename': [
+    1,
+    { ignoreDestructuring: false, ignoreImport: false, ignoreExport: false },
+  ],
+  'no-var': 1,
+  'no-whitespace-before-property': 1,
+  'no-with': 2,
+  'operator-assignment': [1, 'always'],
+  'prefer-const': 1,
+  'quote-props': [1, 'as-needed', { numbers: true }],
+  radix: 1,
+  'require-yield': 1,
+  strict: [1, 'never'],
+  'unicode-bom': [2, 'never'],
+  'unicorn/filename-case': [2, { case: 'kebabCase' }],
+  'unicorn/no-array-instanceof': 1,
+  'unicorn/no-new-buffer': 1,
+  'unicorn/prefer-starts-ends-with': 1,
+  'unicorn/prefer-type-error': 1,
+  'use-isnan': 1,
+  'valid-typeof': 1,
+  'wrap-iife': [1, 'any'],
+  yoda: [1, 'never'],
+}
+
+module.exports = { baseRules }
